+++ app/components/services/information_description_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Services::InformationDescriptionComponent < ApplicationComponent
+  def initialize(order_type:, service_title:, offer:)
+    @order_type = order_type
+    @service_title = service_title
+    @offer = offer
+  end
+end